Improve HelpScreen
The styles were cleaned up and a new component - HelpOption was created
in order to avoid redundancy
